requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=2.0.0
reportlab>=4.0.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
python-dotenv>=1.0.0
dj-database-url>=2.0.0
